requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
mcp[cli]
httpx
